requirements.txt
streamlit
textblob
opencv-python-headless
pyttsx3
python-docx
pandas
deepface
